feat: remove diagram references on diagram removal (COR-000) (#8294)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements VF-XXX**

### Brief description. What is this change?

<!-- Build up some context for your teammates on the changes made here and potential tradeoffs made and/or highlight any topics for discussion -->

copilot:summary

+++ apps/realtime/src/flow/flow.logux.controller.ts
@@ -26,7 +26,9 @@ export class FlowLoguxController {
     @Payload() { data, context }: Actions.Flow.CreateOne.Request,
     @AuthMeta() auth: AuthMetaPayload
   ): Promise<Actions.Flow.CreateOne.Response> {
-    return this.service.createManyAndBroadcast([data], { auth, context }).then(([result]) => ({ data: this.service.toJSON(result), context }));
+    return this.service
+      .createManyAndBroadcast([data], { auth, context })
+      .then(([result]) => ({ data: this.service.toJSON(result), context }));
   }
 
   @Action.Async(Actions.Flow.CreateMany)
@@ -39,7 +41,9 @@ export class FlowLoguxController {
     @Payload() { data, context }: Actions.Flow.CreateMany.Request,
     @AuthMeta() auth: AuthMetaPayload
   ): Promise<Actions.Flow.CreateMany.Response> {
-    return this.service.createManyAndBroadcast(data, { auth, context }).then((results) => ({ data: this.service.mapToJSON(results), context }));
+    return this.service
+      .createManyAndBroadcast(data, { auth, context })
+      .then((results) => ({ data: this.service.mapToJSON(results), context }));
   }
 
   @Action.Async(Actions.Flow.DuplicateOne)
@@ -111,7 +115,7 @@ export class FlowLoguxController {
   @BroadcastOnly()
   @UseRequestContext()
   async deleteOne(@Payload() { id, context }: Actions.Flow.DeleteOne, @AuthMeta() auth: AuthMetaPayload) {
-    const result = await this.service.deleteManyAndSync([id], { context });
+    const result = await this.service.deleteManyAndSync([id], { userID: auth.userID, context });
 
     // overriding entities cause it's broadcasted by decorator
     await this.service.broadcastDeleteMany({ ...result, delete: { ...result.delete, flows: [] } }, { auth, context });
@@ -126,7 +130,7 @@ export class FlowLoguxController {
   @BroadcastOnly()
   @UseRequestContext()
   async deleteMany(@Payload() { ids, context }: Actions.Flow.DeleteMany, @AuthMeta() auth: AuthMetaPayload) {
-    const result = await this.service.deleteManyAndSync(ids, { context });
+    const result = await this.service.deleteManyAndSync(ids, { userID: auth.userID, context });
 
     // overriding entities cause it's broadcasted by decorator
     await this.service.broadcastDeleteMany({ ...result, delete: { ...result.delete, flows: [] } }, { auth, context });
